2015-03-24 13:42 UTC+0100 Przemyslaw Czerpak (druzus/at/poczta.onet.pl)   * include/harbour.hbx   * src/vm/cmdarg.c     + added new PRG function          hb_ACmdLine() -> <array>       borrowed from Viktor's branch (2015-03-12 03:02 UTC+0100 Viktor Szakats)       It returns array with all command line parameters. Unlike           hb_AParams( __dbgProcLevel() - 1 )       hb_ACmdLine() returns all parameters also hidden ones just like       hb_CmdLine() function and hb_argv()/hb_argc()       Please also remember that hb_AParams() returns current value of       parameters which could be changed bu user code so they can be different       then original parameters passed to given function.
  * src/rtl/gtchrmap.c
    * casting

  * include/std.ch
    * synced casing with Viktor's branch

+++ src/rtl/gtchrmap.c
@@ -304,7 +304,7 @@ static int chrmap_parse( FILE * fp, const char * pszTerm, int * nTransTbl, const
                   ++s;
                *s = '\0';
                s = buf;
-               i = strlen( pszTerm );
+               i = ( int ) strlen( pszTerm );
                while( isTerm == 0 && ( s = strstr( s + 1, pszTerm ) ) != NULL )
                {
                   if( *( s - 1 ) == '|' &&
